game_service_api/requirements.txt
flask
pymongo
pillow